+++ Misc/NEWS.d/next/Core and Builtins/2022-11-07-14-16-59.gh-issue-81057.3uKlLQ.rst
@@ -1,0 +1,3 @@
+The 18 global C variables holding the state of the allocators have been
+moved to ``_PyRuntimeState``.  This is a strictly internal change with no
+change in behavior.
